refactored chart to display in order with correct relative durations
